+++ validationframework-swing/src/main/java/com/google/code/validationframework/swing/property/ComponentSizeProperty.java
@@ -29,21 +29,23 @@ import com.google.code.validationframework.api.common.Disposable;
 import com.google.code.validationframework.base.property.AbstractReadableWritableProperty;
 
 import java.awt.Component;
+import java.awt.Dimension;
 import java.awt.event.ComponentEvent;
 import java.awt.event.ComponentListener;
 
 /**
- * Readable/writable property representing the visible state of a {@link Component}.
+ * Readable/writable property representing the size of a {@link Component}.
  * <p/>
- * It is possible to control the visible state of the component by setting the value of this property or by calling the
- * {@link Component#setVisible(boolean)} method of that component.
+ * It is possible to control the size of the component by setting the value of this property or by calling the {@link
+ * Component#setSize(Dimension)} method of that component.
  * <p/>
- * If the value of this property is set to null, the component visible state will not be changed.
+ * However, note that the layout manager of the parent container may also modify the size of the component.
  */
-public class ComponentVisibleProperty extends AbstractReadableWritableProperty<Boolean, Boolean> implements Disposable {
+public class ComponentSizeProperty extends AbstractReadableWritableProperty<Dimension,
+        Dimension> implements Disposable {
 
     /**
-     * Visible state tracker.
+     * Size tracker.
      */
     private class EventAdapter implements ComponentListener {
 
@@ -52,7 +54,7 @@ public class ComponentVisibleProperty extends AbstractReadableWritableProperty<B
          */
         @Override
         public void componentShown(ComponentEvent e) {
-            updateFromComponent();
+            // Nothing to be done
         }
 
         /**
@@ -60,7 +62,7 @@ public class ComponentVisibleProperty extends AbstractReadableWritableProperty<B
          */
         @Override
         public void componentHidden(ComponentEvent e) {
-            updateFromComponent();
+            // Nothing to be done
         }
 
         /**
@@ -68,7 +70,7 @@ public class ComponentVisibleProperty extends AbstractReadableWritableProperty<B
          */
         @Override
         public void componentResized(ComponentEvent e) {
-            // Nothing to be done
+            updateFromComponent();
         }
 
         /**
@@ -80,41 +82,41 @@ public class ComponentVisibleProperty extends AbstractReadableWritableProperty<B
         }
 
         /**
-         * Sets the value of the property based on the visible state of the component.
+         * Sets the value of the property based on the size of the component.
          */
         private void updateFromComponent() {
             updatingFromComponent = true;
-            setValue(component.isVisible());
+            setValue(component.getSize());
             updatingFromComponent = false;
         }
     }
 
     /**
-     * Component to track the visible state for.
+     * Component to track the size of.
      */
     private final Component component;
 
     /**
-     * Visible state tracker.
+     * Size tracker.
      */
     private final EventAdapter eventAdapter = new EventAdapter();
 
     /**
      * Current property value.
      */
-    private Boolean value = null;
+    private Dimension value = null;
 
     /**
-     * Flag indicating whether the {@link #setValue(Boolean)} call is due to a property change event.
+     * Flag indicating whether the {@link #setValue(Dimension)} call is due to a property change event.
      */
     private boolean updatingFromComponent = false;
 
     /**
      * Constructor specifying the component for which the property applies.
      *
-     * @param component Component whose visible property is to be tracked.
+     * @param component Component whose size property is to be tracked.
      */
-    public ComponentVisibleProperty(Component component) {
+    public ComponentSizeProperty(Component component) {
         super();
 
         // Hook to component
@@ -122,7 +124,7 @@ public class ComponentVisibleProperty extends AbstractReadableWritableProperty<B
         this.component.addComponentListener(eventAdapter);
 
         // Set initial value
-        value = component.isVisible();
+        value = component.getSize();
     }
 
     /**
@@ -138,7 +140,7 @@ public class ComponentVisibleProperty extends AbstractReadableWritableProperty<B
      * @see AbstractReadableWritableProperty#getValue()
      */
     @Override
-    public Boolean getValue() {
+    public Dimension getValue() {
         return value;
     }
 
@@ -146,13 +148,13 @@ public class ComponentVisibleProperty extends AbstractReadableWritableProperty<B
      * @see AbstractReadableWritableProperty#setValue(Object)
      */
     @Override
-    public void setValue(Boolean value) {
+    public void setValue(Dimension value) {
         if (updatingFromComponent) {
-            Boolean oldValue = this.value;
+            Dimension oldValue = this.value;
             this.value = value;
             maybeNotifyListeners(oldValue, this.value);
         } else if (value != null) {
-            component.setVisible(value);
+            component.setSize(value);
         }
     }
 }