+++ CHANGES.md
@@ -7,6 +7,7 @@
 * Added convenience method in the FormatTransformer to get the format property
 * Added method in AbstractReadableProperty to get the value change listeners
 * Added JLabel and JButton related properties
+* Added component size and location related properties
 * Added method to the FormatWrapper to get the delegate format
 
 # Version 3.1.0